requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
flask-mysql-connector==1.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
mysql-connector-python==8.0.25
numpy==1.20.3
pandas==1.2.4
protobuf==3.17.0
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
plotly==5.2.2
tenacity==8.0.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
Werkzeug==2.0.1
